Sounds
Added acceleration and collision sounds.  Need to refine the collision
sound a bit.  When there is a wreck, the sound goes off too many times.

Going to add an idle sound next.  For now, the only car noise is the
acceleration when "W" is pressed.  Also need to add code for the xbox
controller left joystick forward.