CarIdle and Game Over Sounds
Added a CarIdle sound when the carts are not accelerating.

Added a game_finish sound when the cart finishes the 3rd lap.

Still need to refine this a tad bit.  You can hear the game over sound
playing in the distance on the other carts.  A minor annoyance but
something I'd like to fix.